improves readme
